requirements.txt
appdirs==1.4.3
appnope==0.1.0
asgiref==3.2.7
astroid==2.3.3
backcall==0.1.0
certifi==2019.11.28
decorator==4.4.2
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==3.0.5
psycopg2==2.8.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.6.1
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
traitlets==4.3.3
virtualenv==20.0.15
virtualenv-clone==0.5.4
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
